refactor: move module query-methods into collection-decorator
