+++ lib/draper.rb
@@ -23,7 +23,7 @@ require 'draper/factory'
 require 'draper/decorated_association'
 require 'draper/helper_support'
 require 'draper/view_context'
-require 'draper/query_methods'
+require 'draper/load_strategy'
 require 'draper/collection_decorator'
 require 'draper/undecorate'
 require 'draper/decorates_assigned'
