+++ spec/draper/load_strategy_spec.rb
